chore(suite): refactoring: split DeviceSelector to two components
